requirements.txt
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Deprecated==1.2.14
Flask==2.3.2
gcloud==0.18.3
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jwcrypto==1.5.0
MarkupSafe==2.1.3
numpy==1.25.1
oauth2client==4.1.3
pandas==2.0.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pyparsing==3.1.0
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
